--- src/lib/radio-group-list.ts
@@ -1,22 +1,0 @@
-type ListSize = 'sm' | 'default' | 'lg';
-type ListDensity = 'default' | 'comfortable';
-
-const verticalLegendMb: Record<ListSize, Record<ListDensity, string>> = {
-  sm: { default: 'mb-3', comfortable: 'mb-4' },
-  default: { default: 'mb-3', comfortable: 'mb-4' },
-  lg: { default: 'mb-4', comfortable: 'mb-5' },
-};
-
-function listLegendMb(
-  size: ListSize,
-  density: ListDensity,
-  orientation: 'vertical' | 'horizontal',
-) {
-  if (orientation === 'horizontal') {
-    return 'mb-3';
-  }
-
-  return verticalLegendMb[size][density];
-}
-
-export { listLegendMb, type ListDensity, type ListSize };
